+++ NEWS.md
@@ -1,5 +1,17 @@
 # Changelog OBIC
 
+## Version 0.11.1 2020-09-02
+### Changed
+* The water holding capacity is now between 0 and 1 (as theta should be) and uses new evaluation function #67
+* Lower range of `D_NLV` is set to -30
+* Upper range of `D_MG` is set to 1000
+* Range of `D_P_DU` is set between 0 and 1
+
+### Fixed
+* Fixed the use of `B_LG_CBS` at sulphur
+* `ID` can now also be character and does not have to be numeric #66
+
+
 ## Version 0.11.0 2020-01-16
 ### Added
 * Adds two other pedotransfer functions (PTFs) to calculate water retention parameters #53